Promenjena struktura od formu da moze da se upotrebi za updejt i kasnije mozda edit
